Modificando função de compra de tickets, resolvendo empecilhos anteriores.
